refactor: remove duplicates and move it to custom hooks
