+++ front/src/pages/User/Home/PublicLessons/PublicLessons.mobile.jsx
@@ -1,74 +1,23 @@
-import { useEffect, useMemo, useState } from 'react';
 import { Col, Row, Skeleton } from 'antd';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
-import { useHistory, useLocation } from 'react-router-dom';
-import { useQuery } from 'react-query';
 import emptyImg from '@sb-ui/resources/img/empty.svg';
 import LessonBlock, { PUBLIC_LESSON } from '@sb-ui/components/LessonBlock';
-import { getPublicLessons } from '@sb-ui/utils/api/v1/student';
-import { getQueryPage } from '@sb-ui/utils/utils';
-import { USER_PUBLIC_LESSONS_BASE_KEY } from '@sb-ui/utils/queries';
-import * as S from './PublicLessons.mobile.styled';
 import { PAGE_SIZE } from './constants';
+import { useLessons } from './useLessons';
+import * as S from './PublicLessons.mobile.styled';
 
 const PublicLessonsMobile = ({ searchLessons }) => {
   const { t } = useTranslation();
-  const location = useLocation();
-  const queryPage = useMemo(() => location.search, [location]);
-  const history = useHistory();
-  const [search, setSearch] = useState(null);
-  const [currentPage, setCurrentPage] = useState(1);
-
   const {
-    data: responseData,
+    data,
+    total,
+    currentPage,
     isLoading,
-    isPreviousData,
     isSuccess,
-  } = useQuery(
-    [
-      USER_PUBLIC_LESSONS_BASE_KEY,
-      {
-        offset: (currentPage - 1) * PAGE_SIZE,
-        limit: PAGE_SIZE,
-        search,
-      },
-    ],
-    getPublicLessons,
-    { keepPreviousData: true },
-  );
-
-  const { data, total } = responseData || {};
-
-  useEffect(() => {
-    if (data?.length === 0 && total !== 0) {
-      setCurrentPage(1);
-      history.replace({
-        search: ``,
-      });
-    }
-  }, [data, history, total]);
-
-  useEffect(() => {
-    const { incorrect, page } = getQueryPage(queryPage);
-    setCurrentPage(page);
-    if (incorrect || page === 1) {
-      history.replace({
-        search: ``,
-      });
-    }
-  }, [history, queryPage]);
-
-  useEffect(() => {
-    setSearch(searchLessons);
-  }, [searchLessons]);
-
-  const onChangeLessonsPage = (page) => {
-    setCurrentPage(page);
-    history.push({
-      search: `?page=${page}`,
-    });
-  };
+    isPreviousData,
+    onChangeLessonsPage,
+  } = useLessons(searchLessons);
 
   if (isSuccess && data?.length === 0) {
     return (
